+++ mountpoint-s3/src/cli.rs
@@ -27,6 +27,7 @@ const LOGGING_OPTIONS_HEADER: &str = "Logging options";
 const CACHING_OPTIONS_HEADER: &str = "Caching options";
 const ADVANCED_OPTIONS_HEADER: &str = "Advanced options";
 
+const MOUNTPOINT_LOG_TARGET: &str = "mountpoint_s3";
 const FSTAB_DOCS: &str = "
 Alternative fstab style:
   mount-s3 <BUCKET> <DIRECTORY> -o <OPTIONS>
@@ -669,7 +670,7 @@ impl CliArgs {
             let mut filter = if self.debug {
                 String::from("debug")
             } else {
-                String::from("info")
+                format!("warn,{MOUNTPOINT_LOG_TARGET}=info")
             };
             let crt_verbosity = if self.debug_crt { "debug" } else { "off" };
             filter.push_str(&format!(",{AWSCRT_LOG_TARGET}={crt_verbosity}"));