+++ mountpoint-s3/CHANGELOG.md
@@ -1,9 +1,17 @@
-## Unreleased (v1.21.0)
+## Unreleased (v1.22.0)
+
+## v1.21.0 (Oct 30, 2025)
+
+### New features
+
+* Mountpoint for Amazon S3 adds support for exporting metrics using OTel Protocol. See [METRICS documentation](https://github.com/awslabs/mountpoint-s3/blob/main/doc/METRICS.md). ([1685](https://github.com/awslabs/mountpoint-s3/pull/1685))
+
+### Other changes
 
 * Change default logging level from WARN to INFO to improve visibility of Mountpoint operational messages. ([#1605](https://github.com/awslabs/mountpoint-s3/pull/1605), [#1668](https://github.com/awslabs/mountpoint-s3/pull/1668))
-* Change FUSE and S3 request metric names in logs ([#1630](https://github.com/awslabs/mountpoint-s3/pull/1630), [#1653](https://github.com/awslabs/mountpoint-s3/pull/1653))
-* Fixed "file does not exist" errors during concurrent directory listing operations ([#1648](https://github.com/awslabs/mountpoint-s3/pull/1648))
-* Change metric logging format to add metric units ([#1677](https://github.com/awslabs/mountpoint-s3/pull/1677))
+* Change FUSE and S3 request metric names in logs. ([#1630](https://github.com/awslabs/mountpoint-s3/pull/1630), [#1653](https://github.com/awslabs/mountpoint-s3/pull/1653))
+* Change metric logging format to add metric units. ([#1677](https://github.com/awslabs/mountpoint-s3/pull/1677))
+* Fixed "file does not exist" errors during concurrent directory listing operations. ([#1648](https://github.com/awslabs/mountpoint-s3/pull/1648))
 
 ## v1.20.0 (Sep 12, 2025)